Rename `RW` to `Mode` and make it part of `lib`.
